a couple of watcher examples + an s/Added/Applied
watcher logic feels a lot nicer to handle from the apps now.
